Move collision logic to hover event + some other adjustments
Moved the collision logic to hover event instead of registering
listeners for better performance.

Added focus triggers (to show the popup) + some styles. The info preview
component needs to wrap the links to be able to handle the focus
logic.

Removed the tooltip role since the information is always on the dom tree
and linked to the card title through aria-labelledby.